requirements.txt
altgraph==0.10.2
azure==0.11.1
bdist-mpkg==0.5.0
bonjour-py==0.3
cffi==1.1.2
cryptography==0.9.3
dj-database-url==0.3.0
Django==1.8.3
django-filter==0.11.0
django-simple-captcha==0.4.6
djangorestframework==3.2.3
dnspython==1.12.0
docutils==0.12
enum34==1.0.4
futures==3.0.3
idna==2.0
ipaddress==1.0.14
macholib==1.5.1
Markdown==2.6.2
matplotlib==1.3.1
mcstatus==2.1
modulegraph==0.10.4
numpy==1.8.0rc1
pbr==1.4.0
Pillow==2.9.0
py2app==0.7.3
pyasn1==0.1.8
pycparser==2.14
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pyserial==2.7
python-dateutil==1.5
pytz==2013.7
requests==2.7.0
scipy==0.13.0b1
six==1.4.1
sqlparse==0.1.16
statistics==1.0.3.5
stevedore==1.7.0
virtualenv==13.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
xattr==0.6.4
zope.interface==4.1.1
